Add minimal DNS helper application
Add a minimal DNS resolver (built for both i386 and amd64) so that we
can stop using the dns-cgi hack.  This minimal DNS resolver uses
libares (c-ares).  It also capable of returning its output in a format
suitable for passing to a Bourne shell "eval".  This utility can also
independently validate the syntax of an IPv4 address or a hostname.

Put the necessary targets in the Makefile to build this.

Declare the "host" target as PHONY, and while we're at it, declare
other .PHONY targets that should have already been there ("clean" and
"install")

Modify host.c to check hostnames with -h