Rewrite stage1 installer to use debconf
Add a set of templates and rewrite the stage 1 installer to use
debconf, for a more pleasing UI, rather than relying on a
poorly-initialized VT5.  The logic is changed a bit, and we ask the
metapackage question here, as well as making it a bit easier to change
advanced options.