Remove legacy cluster network-fixup code
Now that all stage2 installs require a working network at boot, there
is no need to "fixup" the ifupdown config that netcfg generates.

The installer really shouldn't muck with anything on disk.  Networking
stuff specific to PXE installs should be handled in the postinstall
script, if we ever need to do this in the future.